refactor: remove service copy-questions and method moved to user-question service
