+++ src/answers/answers.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Post, Body, Query, ParseIntPipe } from '@nestjs/common';
 import { plainToInstance } from 'class-transformer';
 
-import { CopyTopicQuestionsService } from '../user-questions/copy-topic-question.service';
+import { UserQuestionsService } from '../user-questions/user-questions.service';
 import { CreateAnswerDto } from './dto/create-answer.dto';
 import { OutputCreateOrUpdateDto } from './dto/output-create-or-update.dto';
 import { AnswersService } from './answers.service';
@@ -10,7 +10,7 @@ import { AnswersService } from './answers.service';
 export class AnswersController {
   constructor(
     private readonly answersService: AnswersService,
-    private readonly copyTopicQuestionsService: CopyTopicQuestionsService,
+    private readonly userQuestionsService: UserQuestionsService,
   ) {}
 
   @Post()
@@ -19,7 +19,7 @@ export class AnswersController {
     @Query('userId', ParseIntPipe) userId: number,
     @Query('topicId', ParseIntPipe) topicId: number,
   ) {
-    await this.copyTopicQuestionsService.copyQuestions(topicId, userId);
+    await this.userQuestionsService.copyQuestions(topicId, userId);
 
     const answer = await this.answersService.createOrUpdate(createAnswerDto);
